requirements.txt
algo-ops
easyocr
pytesseract
numpy
natsort
pyspellchecker
opencv-python-headless==4.5.4.60
shapely